increment_3: внедрены fiber и resty вместо стандарной библиотеки, добавлен доступ к чтению метрик
