+++ README.md
@@ -1,1 +1,33 @@
+# GO METRIC
 
+Учебный проект, представляющий из себя клиент-серверное приложение для сбора метрик.
+
+- go 1.21.13
+- fiber 2.*
+- resty
+
+Поддерживаются 2 типа метрик: **gauge** и **counter**
+
+
+###  Agent
+
+Собирает метрики и отправляет на сервер.
+```
+cd cmd/agent
+go build
+./agent
+```
+
+### Server
+
+Принимает метрики от агента, пишет в хранилище, предоставляет интерфейсы чтения метрик:
+```
+cd cmd/server
+go build
+./server
+```
+
+#### Доступ к метрикам
+
+- http://127.0.0.1:8080/ - текущий стэйт хранища, все известные метрики и их значения на данный момент.
+- GET http://127.0.0.1:8080/value/{metric_type}/{metric_name} - текущее состояние конкретной метрики
