+++ internal/common/types.go
@@ -3,7 +3,7 @@ package common
 type Metric struct {
 	Name string
 	Type string
-	Val  interface{}
+	Val  float64
 }
 
 type Metrics map[string]Metric
