--- internal/handlers/constants.go
@@ -1,6 +1,0 @@
-package handlers
-
-const (
-	urlPrefixGauge   = "/update/gauge/"
-	urlPrefixCounter = "/update/counter/"
-)
