+++ internal/server/server.go
@@ -1,34 +1,28 @@
 package server
 
 import (
-	"github.com/i-eliseyev/go-metric/internal/handlers"
+	"github.com/gofiber/fiber/v2"
+	"github.com/gofiber/template/html/v2"
+	"github.com/i-eliseyev/go-metric/internal/routers"
 	"log"
-	"net/http"
-	"time"
 )
 
-const PORT = ":8080"
-
 func StartServer() error {
+	app := fiber.New(
+		fiber.Config{
+			IdleTimeout:  IdleTimeout,
+			ReadTimeout:  ReadTimeout,
+			WriteTimeout: WriteTimeout,
+			Views:        html.New("./internal/templates", ".html"),
+		},
+	)
+	routers.SetupRouters(app)
+	err := app.Listen(Port)
 
-	mux := http.NewServeMux()
-	server := http.Server{
-		Addr:         PORT,
-		Handler:      mux,
-		IdleTimeout:  10 * time.Second,
-		ReadTimeout:  time.Second,
-		WriteTimeout: time.Second,
-	}
-
-	mux.Handle("/update/gauge/", http.HandlerFunc(handlers.HandleGauge))
-	mux.Handle("/update/counter/", http.HandlerFunc(handlers.HandleCounter))
-
-	log.Println("Ready to work!")
-
-	err := server.ListenAndServe()
 	if err != nil {
 		log.Println(err)
 		return err
 	}
 	return nil
+
 }
