+++ internal/storage/interface.go
@@ -1,0 +1,9 @@
+package storage
+
+import "github.com/i-eliseyev/go-metric/internal/common"
+
+type Storage interface {
+	UpdateGauge(m *common.Metric)
+	UpdateCounter(m *common.Metric)
+	GetMetric(name string) (*common.Metric, error)
+}
