+++ internal/templates/index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Metrics</title>
+</head>
+<body>
+<h1>Metrics</h1>
+<table border="1">
+    <tr>
+        <th>Name</th>
+        <th>Type</th>
+        <th>Value</th>
+    </tr>
+    {{range $key, $value := .}}
+    <tr>
+        <td>{{$key}}</td>
+        <td>{{$value.Type}}</td>
+        <td>{{$value.Value}}</td>
+    </tr>
+    {{end}}
+</table>
+</body>
+</html>
